incorporated review comments
