+++ ocs_ci/ocs/ui/page_objects/block_and_file.py
@@ -181,3 +181,21 @@ class BlockAndFile(StorageSystemDetails):
         )
 
         return used, available
+
+    def get_estimated_days_from_consumption_trend(self):
+        """
+        This will get the web element xpaths from views.py and try to get text from those elements
+
+        This will fetch information from DataFoundation>>Storage>>Block and File page>>Consumption trend card
+
+        Returns:
+            tuple: (get_est_days_from_element, get_avg_from_element)
+
+        """
+        get_est_days_from_element = self.get_element_text(
+            self.validation_loc["locate_estimated_days_along_with_value"]
+        )
+        get_avg_from_element = self.get_element_text(
+            self.validation_loc["locate_average_of_storage_consumption"]
+        )
+        return (get_est_days_from_element, get_avg_from_element)
