+++ ocs_ci/ocs/ui/page_objects/storage_system_details.py
@@ -149,21 +149,3 @@ class StorageSystemDetails(StorageSystemTab):
         if not is_present:
             logger.warning("Storage System Details breadcrumb is not present")
         return is_present
-
-    def get_estimated_days_from_consumption_trend(self):
-        """
-        This will get the web element xpaths from views.py and try to get text from those elements
-
-        This will fetch information from DataFoundation>>Storage>>Block and File page>>Consumption trend card
-
-        Returns:
-            tuple: (get_est_days_from_element, get_avg_from_element)
-
-        """
-        get_est_days_from_element = self.get_element_text(
-            self.validation_loc["locate_estimated_days_along_with_value"]
-        )
-        get_avg_from_element = self.get_element_text(
-            self.validation_loc["locate_average_of_storage_consumption"]
-        )
-        return (get_est_days_from_element, get_avg_from_element)
