+++ tests/functional/ui/test_odf_storage_consumption_trend.py
@@ -1,7 +1,5 @@
 import logging
 import math
-import re
-import time
 
 from ocs_ci.framework import config
 from ocs_ci.framework.testlib import (
@@ -12,14 +10,14 @@ from ocs_ci.framework.testlib import (
     black_squad,
     skipif_ocp_version,
     skipif_ocs_version,
+    skipif_mcg_only,
+    skipif_external_mode,
 )
 from ocs_ci.helpers import helpers
 from ocs_ci.helpers.osd_resize import basic_resize_osd
 from ocs_ci.ocs import constants
-from ocs_ci.ocs.cluster import get_used_and_total_capacity_in_gibibytes
 from ocs_ci.ocs.ocp import OCP
 from ocs_ci.ocs.resources.pod import (
-    get_age_of_cluster_in_days,
     get_mgr_pods,
     get_ceph_tools_pod,
     delete_pods,
@@ -27,6 +25,7 @@ from ocs_ci.ocs.resources.pod import (
 )
 from ocs_ci.ocs.resources.storage_cluster import get_storage_size
 from ocs_ci.ocs.ui.validation_ui import ValidationUI
+from ocs_ci.utility.utils import TimeoutSampler
 
 logger = logging.getLogger(__name__)
 
@@ -40,6 +39,8 @@ POD_OBJ = OCP(kind=constants.POD, namespace=config.ENV_DATA["cluster_namespace"]
 @black_squad
 @skipif_ocp_version("<4.17")
 @skipif_ocs_version("<4.17")
+@skipif_mcg_only
+@skipif_external_mode
 class TestConsumptionTrendUI(ManageTest):
     """
     Few test cases in this class requires manual calculations for 'average of storage consumption'
@@ -56,67 +57,18 @@ class TestConsumptionTrendUI(ManageTest):
 
     """
 
-    def get_est_days_from_ui(self):
-        """
-        Get the value of 'Estimated days until full' from the UI
-
-        Returns:
-            est_days: (int)
-
-        """
-        validation_ui_obj = ValidationUI()
-        collected_tpl_of_days_and_avg = (
-            validation_ui_obj.odf_storagesystems_consumption_trend()
-        )
-        est_days = re.search(r"\d+", collected_tpl_of_days_and_avg[0]).group()
-        logger.info(f"'Estimated days until full' from the UI : {est_days}")
-        return int(est_days)
-
-    def get_avg_consumption_from_ui(self):
-        """
-        Get the value of 'Average storage consumption' from the UI
-
-        Returns:
-            average: (float)
-
-        """
-        validation_ui_obj = ValidationUI()
-        collected_tpl_of_days_and_avg = (
-            validation_ui_obj.odf_storagesystems_consumption_trend()
-        )
-        average = float(
-            re.search(r"-?\d+\.*\d*", collected_tpl_of_days_and_avg[1]).group()
-        )
-        logger.info(f"'Average of storage consumption per day' from the UI : {average}")
-        return average
-
-    def calculate_est_days_and_average_manually(self):
-        """
-        Calculates the 'Estimated days until full' manually by:
-        1. Get the age of the cluster in days
-        2. Get used capacity of the cluster
-        3. Get total capacity of the cluster
-        4. Calculate average consumption of the storage per day
-        5. Calculate the 'Estimated days until full' by using average and available capacity.
-
-        Returns:
-            estimated_days_calculated: (tuple)
-
-        """
-        number_of_days = get_age_of_cluster_in_days()
-        logger.info(f"Age of the cluster in days: {number_of_days}")
-        tpl_of_used_and_total_capacity = get_used_and_total_capacity_in_gibibytes()
-        used_capacity = tpl_of_used_and_total_capacity[0]
-        logger.info(f"The used capacity from the cluster is: {used_capacity}")
-        total_capacity = tpl_of_used_and_total_capacity[1]
-        available_capacity = total_capacity - used_capacity
-        logger.info(f"The available capacity from the cluster is: {available_capacity}")
-        average = used_capacity / number_of_days
-        logger.info(f"Average of storage consumption per day: {average}")
-        estimated_days_calculated = available_capacity / average
-        manual_est = math.floor(estimated_days_calculated)
-        logger.info(f"Estimated days calculated are {manual_est}")
-        return (manual_est, average)
+    def get_active_mgr(self):
+        logger.info("Get mgr pods objs")
+        mgr_objs = get_mgr_pods()
+        toolbox = get_ceph_tools_pod()
+        active_mgr_pod_output = toolbox.exec_cmd_on_pod("ceph mgr stat")
+        active_mgr_pod_suffix = active_mgr_pod_output.get("active_name")
+        logger.info(f"The active MGR pod is {active_mgr_pod_suffix}")
+        active_mgr_deployment_name = "rook-ceph-mgr-" + active_mgr_pod_suffix
+        for obj in mgr_objs:
+            if active_mgr_pod_suffix in obj.name:
+                active_mgr_pod = obj.name
+        return (active_mgr_deployment_name, active_mgr_pod)
 
     def test_consumption_trend_card_ui(self, setup_ui_class):
         """
@@ -152,10 +104,11 @@ class TestConsumptionTrendUI(ManageTest):
             2. Compare the above calculated value with the ‘Estimated days until full’ in the widget (UI
 
         """
-        est_days = self.get_est_days_from_ui()
-        average = self.get_avg_consumption_from_ui()
+        validation_ui_obj = ValidationUI()
+        est_days = validation_ui_obj.get_est_days_from_ui()
+        average = validation_ui_obj.get_avg_consumption_from_ui()
         logger.info(f"From the UI, Estimated Days: {est_days} and Average: {average}")
-        days_avg_tpl = self.calculate_est_days_and_average_manually()
+        days_avg_tpl = validation_ui_obj.calculate_est_days_and_average_manually()
         first_validation = est_days == days_avg_tpl[0]
         # rel_tol 0.1 means upto 10% tolerance, which means that difference between  manual and UI est days is up to 10%
         second_validation = math.isclose(est_days, days_avg_tpl[0], rel_tol=0.1)
@@ -179,9 +132,10 @@ class TestConsumptionTrendUI(ManageTest):
             2. Compare the above calculated value with the ‘Average’ in the widget (UI)
 
         """
-        average = self.get_avg_consumption_from_ui()
+        validation_ui_obj = ValidationUI()
+        average = validation_ui_obj.get_avg_consumption_from_ui()
         logger.info(f"From the UI, Average: {average}")
-        days_avg_tpl = self.calculate_est_days_and_average_manually()
+        days_avg_tpl = validation_ui_obj.calculate_est_days_and_average_manually()
         first_validation = average == days_avg_tpl[1]
         # rel_tol 0.1 means upto 10% tolerance, which means that difference between  manual and UI est days is up to 10%
         second_validation = math.isclose(average, days_avg_tpl[1], rel_tol=0.1)
@@ -204,27 +158,29 @@ class TestConsumptionTrendUI(ManageTest):
             2. Test storage consumption trend from UI is accurate after mgr failover.
 
         """
-        logger.info("Get mgr pods objs")
-        mgr_objs = get_mgr_pods()
-        toolbox = get_ceph_tools_pod()
-        active_mgr_pod_output = toolbox.exec_cmd_on_pod("ceph mgr stat")
-        active_mgr_pod_suffix = active_mgr_pod_output.get("active_name")
-        logger.info(f"The active MGR pod is {active_mgr_pod_suffix}")
-        active_mgr_deployment_name = "rook-ceph-mgr-" + active_mgr_pod_suffix
-        for obj in mgr_objs:
-            if active_mgr_pod_suffix in obj.name:
-                active_mgr_pod = obj.name
+        validation_ui_obj = ValidationUI()
+        (
+            active_mgr_deployment_name_before_failover,
+            active_mgr_pod_before_failover,
+        ) = self.get_active_mgr()
 
-        logger.info(f"Scale down {active_mgr_deployment_name} to 0")
+        logger.info(f"Scale down {active_mgr_deployment_name_before_failover} to 0")
         helpers.modify_deployment_replica_count(
-            deployment_name=active_mgr_deployment_name, replica_count=0
+            deployment_name=active_mgr_deployment_name_before_failover, replica_count=0
         )
-        POD_OBJ.wait_for_delete(resource_name=active_mgr_pod)
+        POD_OBJ.wait_for_delete(resource_name=active_mgr_pod_before_failover)
         # Below sleep is mandatory for mgr failover, if not the same pod will become active again.
-        time.sleep(60)
-        logger.info(f"Scale down {active_mgr_deployment_name} to 1")
+
+        check_failover = (
+            lambda: logger.info("Mgr Failover succeed")
+            if active_mgr_deployment_name_before_failover != self.get_active_mgr()[0]
+            else None
+        )
+
+        TimeoutSampler(timeout=120, sleep=15, func=check_failover)
+        logger.info(f"Scale down {active_mgr_deployment_name_before_failover} to 1")
         helpers.modify_deployment_replica_count(
-            deployment_name=active_mgr_deployment_name, replica_count=1
+            deployment_name=active_mgr_deployment_name_before_failover, replica_count=1
         )
         assert (
             len(get_mgr_pods()) == 2
@@ -232,10 +188,10 @@ class TestConsumptionTrendUI(ManageTest):
         logger.info("Mgr failovered successfully")
 
         logger.info("Now the testing will begin for consumption trend UI")
-        est_days = self.get_est_days_from_ui()
-        average = self.get_avg_consumption_from_ui()
+        est_days = validation_ui_obj.get_est_days_from_ui()
+        average = validation_ui_obj.get_avg_consumption_from_ui()
         logger.info(f"From the UI, Estimated Days: {est_days} and Average: {average}")
-        days_avg_tpl = self.calculate_est_days_and_average_manually()
+        days_avg_tpl = validation_ui_obj.calculate_est_days_and_average_manually()
         first_validation = est_days == days_avg_tpl[0]
         # rel_tol 0.1 means upto 10% tolerance, which means that difference between  manual and UI est days is up to 10%
         second_validation = math.isclose(est_days, days_avg_tpl[0], rel_tol=0.1)
@@ -262,24 +218,33 @@ class TestConsumptionTrendUI(ManageTest):
             4. Should show close to the values before deleting the prometheus pod
 
         """
+        validation_ui_obj = ValidationUI()
         logger.info("Get the value of 'Estimated days until full' from UI")
-        est_days_before = self.get_est_days_from_ui()
+        est_days_before = validation_ui_obj.get_est_days_from_ui()
         logger.info(
             f"The value of 'Estimated days until full' from UI is {est_days_before} before failing prometheus"
         )
-        average_before = self.get_avg_consumption_from_ui()
+        average_before = validation_ui_obj.get_avg_consumption_from_ui()
         logger.info(
             f"'Average of storage consumption per day' from UI is {average_before} before failing prometheus"
         )
         logger.info("Bring down the prometheus")
         list_of_prometheus_pod_obj = get_prometheus_pods()
         delete_pods(list_of_prometheus_pod_obj)
-        time.sleep(120)
-        est_days_after = self.get_est_days_from_ui()
+        est_days_after = ""
+        for sampler in TimeoutSampler(
+            timeout=300, sleep=30, func=validation_ui_obj.get_est_days_from_ui
+        ):
+            if sampler:
+                est_days_after = sampler
+                break
+            else:
+                logger.info("dashboard is not ready yet")
+        est_days_after = validation_ui_obj.get_est_days_from_ui()
         logger.info(
             f"From the UI, Estimated Days: {est_days_after} after prometheus recovered from failure"
         )
-        average_after = self.get_avg_consumption_from_ui()
+        average_after = validation_ui_obj.get_avg_consumption_from_ui()
         logger.info(
             f"'Average of storage consumption' from UI is {average_after} after prometheus recovered from failure"
         )
@@ -300,15 +265,16 @@ class TestConsumptionTrendUI(ManageTest):
         4. 'Estimated days until full' value in the UI, should increase after OSD resize.
 
         """
+        validation_ui_obj = ValidationUI()
         logger.info("Get the value of 'Estimated days until full' from UI")
-        est_days_before = self.get_est_days_from_ui()
+        est_days_before = validation_ui_obj.get_est_days_from_ui()
         logger.info("Performing OSD resize")
         basic_resize_osd(get_storage_size())
         logger.info("After OSD resize, checking consumption trend UI")
         logger.info(
             "Get the value of 'Estimated days until full' from UI after OSD resize"
         )
-        est_days_after = self.get_est_days_from_ui()
+        est_days_after = validation_ui_obj.get_est_days_from_ui()
         assert (
             est_days_after > est_days_before
         ), f"'Estimated days until full' {est_days_after} did not increase after OSD resize"
